feat: allow flight client passthru args
This change allows the caller to pass arguments through to the
underlying flight library when creating a client.  I am bumping the
version number to mark when this feature is added.  Additionally,
to get this to build on my Ubuntu 23.04 host, I needed to add some
version qualifiers to the dependencies which also speeds up the
process as well.