chore: update readme guidance
The folks supporting the client libraries said that there isn't a
priority on updating influxdb3-python to adhere to Alchemy/DB API 2
and requested this change in response to some inquiries about supporting
integration with Apache SuperSet.

This change notes that this library should continue to work in that capacity,
but does not imply any level of support.